:truck: 최근 이벤트, 히스토리 파일 분리
